moved definitions back to xrsdkit.__init__, moved System constructors to xrsdkit.system.__init__
